:package: Move dependencies to pyproject.toml
