+++ pyproject.toml
@@ -1,7 +1,35 @@
-[tool.isort]
-profile = "black"
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-[tool.pytest.ini_options]
-pythonpath = [
-  "src"
+[tool.hatch.build.targets.wheel]
+    packages = ["src/**/*"]
+
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[project]
+name = "mlex_dlsia"
+version = "0.2.0"
+description = "DLSIA models for segmentation in MLExchange"
+readme = { file = "README.md", content-type = "text/markdown" }
+requires-python = ">=3.11"
+
+dependencies = [
+  "qlty==0.2.0",
+  "dvclive==3.44.0",
+  "dlsia==0.3.1",
+  "pydantic>=1.10.15",
+  "tiled[client]==0.2.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "black==24.2.0",
+    "flake8==7.0.0",
+    "isort==5.13.2",
+    "pre-commit==3.6.2",
+    "tiled[all]==0.2.0",
+    "pytest==8.1.1",
+    "pytest-asyncio"
 ]
