--- requirements-dev.txt
@@ -1,7 +1,0 @@
-black==24.3.0
-flake8==7.0.0
-isort==5.13.2
-pre-commit==3.6.2
-tiled[all]==0.1.0a114
-pytest
-starlette==0.38.2
